requirements.txt
app-version==1.0.1
appdirs==1.4.3
attrs==17.3.0
Automat==0.6.0
bootstrap-admin==0.3.9
bottle==0.12.13
cached-property==1.3.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.9
constantly==15.1.0
decorator==4.1.2
defusedxml==0.5.0
Django==1.11
django-autocomplete-light==3.2.10
django-bootstrap-breadcrumbs==0.9.1
django-chartit==0.2.9
django-ckeditor==5.4.0
django-common-helpers==0.9.1
django-cron==0.5.0
django-filter==1.1.0
django-jalali==2.4.5
django-js-asset==1.0.0
django-jsonfield==1.0.1
django-material==1.2.2
django-notify-x==0.1.9
django-report-builder==4.0.3
django-simple-history==1.9.0
djangorestframework==3.4.3
et-xmlfile==1.0.1
Flask==0.12.2
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipython==6.2.1
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==0.24
jdatetime==1.9.0
jdcal==1.3
jedi==0.11.1
Jinja2==2.10
lxml==4.1.1
Markdown==2.6.11
MarkupSafe==1.0
openpyxl==2.5.0
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
requests-toolbelt==0.8.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
Twisted==17.9.0
typing==3.6.2
urllib3==1.22
validate-email==1.3
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.12.2
zeep==2.5.0
zope.interface==4.4.3
